+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Elasticsearch: force `widget="json"` on `body` so the form is properly filled when updating a data source
+
 ## [6.1.0] 2024-03-13
 
 ### Changed